remove old repo VC and replace with paging
