+++ Classes/Repository/RepositoryIssuesViewController.swift
@@ -51,6 +51,9 @@ LoadMoreSectionControllerDelegate {
 
         feed.viewDidLoad()
         feed.adapter.dataSource = self
+        if #available(iOS 11.0, *) {
+            feed.collectionView.contentInsetAdjustmentBehavior = .never
+        }
     }
 
     override func viewWillAppear(_ animated: Bool) {
