+++ Classes/Repository/RepositoryViewController.swift
@@ -2,43 +2,32 @@
 //  RepositoryViewController.swift
 //  Freetime
 //
-//  Created by Sherlock, James on 29/07/2017.
+//  Created by Ryan Nystrom on 9/20/17.
 //  Copyright © 2017 Ryan Nystrom. All rights reserved.
 //
 
 import UIKit
-import IGListKit
+import Tabman
+import Pageboy
 import TUSafariActivity
 import SafariServices
 
-struct RepositoryDetails {
-    let owner: String
-    let name: String
-    let hasIssuesEnabled: Bool
-}
-
-class RepositoryViewController: UIViewController,
-    FeedDelegate,
-    ListAdapterDataSource,
-    SegmentedControlSectionControllerDelegate,
-    LoadMoreSectionControllerDelegate,
-PrimaryViewController {
+class RepositoryViewController: TabmanViewController, PageboyViewControllerDataSource {
 
     private let repo: RepositoryDetails
-    private let client: RepositoryClient
-    private lazy var feed: Feed = { Feed(viewController: self, delegate: self) }()
-
-    private let noReadmeResultsKey = "noReadmeResultsKey" as ListDiffable
-    private let noIssuesResultsKey = "noIssuesResultsKey" as ListDiffable
-    private let noPullRequestsResultsKey = "noPullRequestsResultsKey" as ListDiffable
-
-    private let dataSource: RepositoryDataSource
+    private let client: GithubClient
+    private let controllers: [UIViewController]
 
     init(client: GithubClient, repo: RepositoryDetails) {
         self.repo = repo
-        self.client = RepositoryClient(githubClient: client, owner: repo.owner, name: repo.name)
-        self.dataSource = RepositoryDataSource(hasIssuesEnabled: repo.hasIssuesEnabled)
+        self.client = client
+        self.controllers = [
+            RepositoryOverviewViewController(client: client, repo: repo),
+            RepositoryIssuesViewController(client: client, repo: repo, type: .issues),
+            RepositoryIssuesViewController(client: client, repo: repo, type: .pullRequests),
+        ]
         super.init(nibName: nil, bundle: nil)
+        title = "\(repo.owner)/\(repo.name)"
     }
 
     required init?(coder aDecoder: NSCoder) {
@@ -50,10 +39,15 @@ PrimaryViewController {
 
         navigationItem.backBarButtonItem = UIBarButtonItem(title: " ", style: .plain, target: nil, action: nil)
 
-        feed.viewDidLoad()
-        feed.adapter.dataSource = self
-        title = "\(repo.owner)/\(repo.name)"
-        
+        dataSource = self
+        bar.items = controllers.map { Item(title: $0.title ?? "" ) }
+        bar.appearance = TabmanBar.Appearance({ appearance in
+            appearance.text.font = Styles.Fonts.button
+            appearance.state.color = Styles.Colors.Gray.light.color
+            appearance.state.selectedColor = Styles.Colors.Blue.medium.color
+            appearance.indicator.color = Styles.Colors.Blue.medium.color
+        })
+
         let rightItem = UIBarButtonItem(
             image: UIImage(named: "bullets-hollow"),
             style: .plain,
@@ -64,26 +58,12 @@ PrimaryViewController {
         navigationItem.rightBarButtonItem = rightItem
     }
 
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        rz_smoothlyDeselectRows(collectionView: feed.collectionView)
-    }
-
-    override func viewWillLayoutSubviews() {
-        super.viewWillLayoutSubviews()
-        feed.viewWillLayoutSubviews(view: view)
-    }
-    
     // MARK: Private API
-    
+
     var repoUrl: URL {
         return URL(string: "https://github.com/\(repo.owner)/\(repo.name)")!
     }
-    
-    var repoOwnerUrl: URL {
-        return URL(string: "https://github.com/\(repo.owner)")!
-    }
-    
+
     func shareAction(sender: UIBarButtonItem) -> UIAlertAction {
         return UIAlertAction(title: NSLocalizedString("Send To", comment: ""), style: .default) { [weak self] _ in
             guard let strongSelf = self else { return }
@@ -93,7 +73,7 @@ PrimaryViewController {
             strongSelf.present(controller, animated: true)
         }
     }
-    
+
     func safariAction() -> UIAlertAction {
         return UIAlertAction(title: NSLocalizedString("Open in Safari", comment: ""), style: .default) { [weak self] _ in
             guard let strongSelf = self else { return }
@@ -101,19 +81,17 @@ PrimaryViewController {
             strongSelf.present(controller, animated: true)
         }
     }
-    
+
     func viewOwnerAction() -> UIAlertAction {
         return UIAlertAction(title: NSLocalizedString("View Owner's Profile", comment: ""), style: .default) { [weak self] _ in
             guard let strongSelf = self else { return }
-            let controller = SFSafariViewController(url: strongSelf.repoOwnerUrl)
+            let controller = SFSafariViewController(url: URL(string: "https://github.com/\(strongSelf.repo.owner)")!)
             strongSelf.present(controller, animated: true)
         }
-
     }
-    
+
     func onMore(sender: UIBarButtonItem) {
         let alert = UIAlertController()
-        
         alert.addAction(shareAction(sender: sender))
         alert.addAction(safariAction())
         alert.addAction(viewOwnerAction())
@@ -121,152 +99,18 @@ PrimaryViewController {
         present(alert, animated: true)
     }
 
-    // MARK: Data Loading/Paging
+    // MARK: PageboyViewControllerDataSource
 
-    private func update(dismissRefresh: Bool, animated: Bool) {
-        feed.finishLoading(dismissRefresh: dismissRefresh, animated: animated)
+    func numberOfViewControllers(in pageboyViewController: PageboyViewController) -> Int {
+        return controllers.count
     }
 
-    func reload() {
-        client.fetchReadme { result in
-            switch result {
-            case .error: break
-            case .success(let readme):
-                self.dataSource.setReadme(
-                    readme,
-                    width: self.view.bounds.width,
-                    owner: self.repo.owner,
-                    repo: self.repo.name,
-                    completion: {
-                    self.update(dismissRefresh: true, animated: true)
-                })
-            }
-        }
-
-//        client.load(containerWidth: view.bounds.width) { result in
-//            switch result {
-//            case .error: break
-//            case .success(let payload):
-//                self.dataSource.reset(
-//                    issues: payload.issues.models,
-//                    issuesNextPage: payload.issues.nextPage,
-//                    pullRequests: payload.pullRequests.models,
-//                    pullRequestsNextPage: payload.pullRequests.nextPage
-//                )
-//                self.update(dismissRefresh: true, animated: true)
-//            }
-//        }
+    func viewController(for pageboyViewController: PageboyViewController, at index: PageboyViewController.PageIndex) -> UIViewController? {
+        return controllers[index]
     }
 
-    func loadNextPage() {
-        let width = view.bounds.width
-
-        switch dataSource.state {
-        case .readme: return
-        case .issues:
-            client.loadIssues(nextPage: dataSource.issuesNextPage, containerWidth: width, completion: { result in
-                switch result {
-                case .error: break
-                case .success(let payload):
-                    self.dataSource.appendIssues(issues: payload.models, page: payload.nextPage)
-                    self.update(dismissRefresh: true, animated: false)
-                }
-            })
-        case .pullRequests:
-            client.loadPullRequests(nextPage: dataSource.pullRequestsNextPage, containerWidth: width, completion: { result in
-                switch result {
-                case .error: break
-                case .success(let payload):
-                    self.dataSource.appendPullRequests(pullRequests: payload.models, page: payload.nextPage)
-                    self.update(dismissRefresh: true, animated: false)
-                }
-            })
-        }
-    }
-
-    // MARK: FeedDelegate
-
-    func loadFromNetwork(feed: Feed) {
-        reload()
-    }
-
-    func loadNextPage(feed: Feed) -> Bool {
-        return false
-    }
-
-    // MARK: ListAdapterDataSource
-
-    func objects(for listAdapter: ListAdapter) -> [ListDiffable] {
-        var builder = [ListDiffable]()
-        builder.append(dataSource.selection)
-
-        let models = dataSource.selectionModels
-        builder += models
-
-        if models.count == 0, feed.status == .idle {
-            switch dataSource.state {
-            case .readme: builder.append(noReadmeResultsKey)
-            case .issues: builder.append(noIssuesResultsKey)
-            case .pullRequests: builder.append(noPullRequestsResultsKey)
-            }
-        }
-
-        return builder
-    }
-
-    func listAdapter(_ listAdapter: ListAdapter, sectionControllerFor object: Any) -> ListSectionController {
-        guard let object = object as? ListDiffable else { fatalError("Object does not conform to ListDiffable") }
-
-        // 28 is the default height of UISegmentedControl
-        let controlHeight = 28 + 2*Styles.Sizes.rowSpacing
-
-        if object === noReadmeResultsKey {
-            return RepositoryEmptyResultsSectionController(
-                topInset: controlHeight,
-                topLayoutGuide: topLayoutGuide,
-                bottomLayoutGuide: bottomLayoutGuide,
-                type: .readme
-            )
-        } else if object === noIssuesResultsKey {
-            return RepositoryEmptyResultsSectionController(
-                topInset: controlHeight,
-                topLayoutGuide: topLayoutGuide,
-                bottomLayoutGuide: bottomLayoutGuide,
-                type: .issues
-            )
-        } else if object === noPullRequestsResultsKey {
-            return RepositoryEmptyResultsSectionController(
-                topInset: controlHeight,
-                topLayoutGuide: topLayoutGuide,
-                bottomLayoutGuide: bottomLayoutGuide,
-                type: .pullRequests
-            )
-        } else if object === dataSource.selection {
-            return SegmentedControlSectionController(delegate: self, height: controlHeight)
-        } else if object === dataSource.loadMore {
-            return LoadMoreSectionController(delegate: self)
-        } else if object is RepositoryIssueSummaryModel {
-            return RepositorySummarySectionController(client: client.githubClient, repo: repo)
-        } else if object is RepositoryReadmeModel {
-            return RepositoryReadmeSectionController()
-        }
-
-        fatalError("Could not find section controller for object")
-    }
-
-    func emptyView(for listAdapter: ListAdapter) -> UIView? {
+    func defaultPage(for pageboyViewController: PageboyViewController) -> PageboyViewController.Page? {
         return nil
     }
 
-    // MARK: SegmentedControlSectionControllerDelegate
-
-    func didChangeSelection(sectionController: SegmentedControlSectionController, model: SegmentedControlModel) {
-        update(dismissRefresh: false, animated: false)
-    }
-
-    // MARK: LoadMoreSectionControllerDelegate
-
-    func didSelect(sectionController: LoadMoreSectionController) {
-        loadNextPage()
-    }
 }
